fix cronjob command
I couldn't get the `...plugin-janitor/cleancache/...` cronjob to run. It always returned a 404 error. But when I changed it to any job name found in the plugins index.php file such as `...plugin-janitor/clear/...` or `..clean` or `..flush`, it returned a successful 200. I'm thinking the `cleancache` command via cron wasn't right and suggest an update to the docs.